Adding access to home directory
Adresses #4.

Merge request #5 was closed by accident (deleting of personal repo), so I re-opened it. Please see my last comment on #5 for reasons to allow access to complete folder.

In addition to my previous change this variant includes a proposed fix to disallow access to hidden folders and files (refer to cli commands on https://askubuntu.com/questions/1086529/how-to-give-a-flatpak-app-access-to-a-directory). Maybe someone can check the syntax.